Create two instances of the K3OSConfig CR reconciler
One handles duties with leader election enabled so it can update
K3OSConfig CRs and the other one is going to handle updating the
Node object on each instance running inside the cluster for the
node it’s running on.

After getting started with creating another manager, 2 metrics servers
and fighting with the whole thing only to find out that the controller
manager can actually handle this was a great learning experience.
The code is so clean (IMHO) and it’s actually not a lot to make this
work.