Jump through several hoops in order to make the feature added in 9c6eccd finally work
Just wow. 2 wrappers were involved in allowing a regular controller
that’s not a webhook server to be executed when the operator instance
isn’t the leader. We still need a leader to handle updating a particular
K3OSConfig CR but to update each Node object the controller must be
executed on all instances.