Wrap reconcile error that’s returned
This is another approach to handle terminal errors in a reconciler.
For the case where an object was deleted (like the node, the secret or
the K3OSConfig CR object) we have to return a nil error to the caller
so that our reconciler isn’t called anymore. I might flesh this out
more in the future with some functional options to configure the
response error further. For now, this will do.