Add a PodSecurityPolicy for the upcoming node config management feature
We want to update the K3OS config.yaml on each k3OS cluster node which
must be mounted via host path. To harden the operator this policy
should be as restrictive as possible (still a ways to go).

For now it makes sure:
- container is run as non-root
- only allows host path mounts starting at /var/lib/rancher/k3os
- no host network / IPC / PID capability
- must drop ALL capabilities
- cannot escalate any privileges and cannot be run privileged

If the operator evolves we might need to enable some of these features
again.