Make PodSecurityPolicy work when the admission plugin is enabled
It takes some work to get it running on k3OS but I managed. The policy
needed to be fixed and the operator actually needs to write to the
/var/lib/rancher/k3os/config.yaml file which is owned by root which
required switching to a root based image. For now, this is good enough
and the policy isn’t very permissive but yes, it could be better.