Move to go-flags library for configuration management
This completely replaces the configuration management that was kept
in various places with constants for environment variable names and
now there’s only once place to look for the configuration.
The configuration is now also strictly passed to the components that
need it so there’s no global config struct anywhere.